+++ apps/alpha-test-app/CHANGELOG.md
@@ -1,5 +1,33 @@
 # alpha-test-app
 
+## 0.0.1-alpha.6
+
+### Patch Changes
+
+- Updated dependencies [eda3982]
+  - sit-onyx@0.1.0-alpha.9
+
+## 0.0.1-alpha.5
+
+### Patch Changes
+
+- Updated dependencies [c62476d]
+  - sit-onyx@0.1.0-alpha.8
+
+## 0.0.1-alpha.4
+
+### Patch Changes
+
+- Updated dependencies [fdada3a]
+  - sit-onyx@0.1.0-alpha.7
+
+## 0.0.1-alpha.3
+
+### Patch Changes
+
+- Updated dependencies [bd5040f]
+  - sit-onyx@0.1.0-alpha.6
+
 ## 0.0.1-alpha.2
 
 ### Patch Changes