+++ packages/sit-onyx/CHANGELOG.md
@@ -1,5 +1,29 @@
 # sit-onyx
 
+## 0.1.0-alpha.9
+
+### Minor Changes
+
+- eda3982: feat(OnyxCheckbox): add property `disabled`
+
+## 0.1.0-alpha.8
+
+### Patch Changes
+
+- c62476d: fix(OnyxIcon): remove default browser margin
+
+## 0.1.0-alpha.7
+
+### Minor Changes
+
+- fdada3a: Implemented barebone RadioButtonGroup and RadioButton support component
+
+## 0.1.0-alpha.6
+
+### Minor Changes
+
+- bd5040f: feat: add `OnyxIcon` component
+
 ## 0.1.0-alpha.5
 
 ### Minor Changes